fix(web): issues with auth due to crfs tokens
