+++ .env.example
@@ -17,6 +17,13 @@ API_BASE_URL=http://localhost:50051
 # AUTH_SERVICE_URL=https://www.eurora-labs.com
 # API_BASE_URL=https://api.eurora-labs.com
 
+# Browser origins the SPA is served from. Drives CORS, the origin
+# guard middleware, and cookie-mode auth detection. Comma-separated.
+WEB_ALLOWED_ORIGINS=http://localhost:5173
+# Set to false in dev so cookies are issued without the Secure flag,
+# which the browser would otherwise drop on plain HTTP.
+AUTH_COOKIE_SECURE=false
+
 APPROVED_EMAILS=*
 ASSET_STORAGE_BACKEND=fs
 ASSET_STORAGE_FS_ROOT=../assets
